requirements.txt
Flask~=1.1.2
PyJWT~=1.7.1
Werkzeug~=1.0.1
flask_sqlalchemy==2.4.1
